[check.c] Organized check-element related function
